Rename RPC files to match primary exports
File names should reflect what they export. The Connect-era names
no longer described their contents after the transcoder migration.
