+++ javascript/packages/rpc/services.ts
@@ -1,5 +1,6 @@
 import { create, createRegistry, fromJson, toJson } from '@bufbuild/protobuf';
 
+import { createFetchTransport } from './create-fetch-transport';
 import { TypedStructSchema } from './gen/michelangelo/api/typed_struct_pb';
 import { DeploymentService } from './gen/michelangelo/api/v2/deployment_svc_pb';
 import { InferenceServerService } from './gen/michelangelo/api/v2/inference_server_svc_pb';
@@ -9,7 +10,6 @@ import { PipelineService } from './gen/michelangelo/api/v2/pipeline_svc_pb';
 import { ProjectService } from './gen/michelangelo/api/v2/project_svc_pb';
 import { TriggerRunService } from './gen/michelangelo/api/v2/trigger_run_svc_pb';
 import { getRuntimeConfig } from './runtime-config';
-import { createFetchTransport } from './transport';
 
 import type { DescService, JsonValue } from '@bufbuild/protobuf';
 import type { FetchTransport, ServiceClient, Services } from './types';
